Add TableCaption/Entity to Telemetry Logs (#66)
* Add TableCaption/Entity to Telemetry Logs

At the moment the Table Caption is only used in the Start-Export Event. 
We have no solution to see when the export for a given table started and stopped. With this change you can:
1. See when the export for a given Table started
2. See how many datasets to export are found
3.. See when the export for a given Table finished

This way you can monitor the mean export time per Table and can react to slower exports or find missing SystemModifiedAt keys.

* Add more detailed Log to ExportTableDeletes

* Increase App version

* Fix EntityCount Rec in ExportTableDeletes